Get back to minpac
I am a little tired of dealing with the bugs caused by the new features added to jetpack.
I like the performance aspect, but it seems that jetpack is trying to improve other aspects
as well, which makes the plugin a bit more fragile (e.g. lazy loading).
So I have decided to go back to minpac which is simpler and more stable.